feature: Add TryCollect and TryFromIterator.
