+++ src/lib.rs
@@ -257,11 +257,10 @@
 //!
 //! Many thanks for these dictionaries and tools.
 
-pub mod clone;
 pub mod inc_search;
 mod internal_data_structure;
 pub mod map;
-mod trie;
-pub mod try_from_iterator;
+pub mod trie;
+pub mod try_collect;
 // pub use try_from_iterator::TryFromIterator;
-pub use clone::{Trie, TrieBuilder};
+pub use trie::clone::{Trie, TrieBuilder};
