+++ src/trie/clone.rs
@@ -1,48 +1,56 @@
-use super::inc_search::IncSearch;
-use super::map::{self, PostfixIter, PrefixIter, SearchIter};
-use frayed::Defray;
+use crate::inc_search::IncSearch;
+use super::map::{self};
 
-pub struct Trie<Label> {
-    inner: map::Trie<Label, ()>,
+use derive_deref::*;
+use std::clone::Clone;
+
+#[derive(Deref, DerefMut)]
+pub struct Trie<Label: Clone> {
+    pub inner: map::Trie<Label, ()>,
 }
-pub struct TrieBuilder<Label> {
+pub struct TrieBuilder<Label: Clone> {
     inner: map::TrieBuilder<Label, ()>,
 }
 
-impl<Label: Ord> Trie<Label> {
+impl<Label: Ord + Clone> Trie<Label> {
     /// Return true if `query` is an exact match.
     pub fn exact_match(&self, query: impl AsRef<[Label]>) -> bool {
         self.inner.exact_match(query).is_some()
     }
 
-    /// Return the common prefixes of `query`.
-    pub fn common_prefix_search<Query>(
-        &self,
-        query: Query,
-    ) -> Defray<PrefixIter<'_, Label, (), Query>>
+    /// Return the common prefixes of `query`, cloned.
+    pub fn common_prefix_search(&self, query: impl AsRef<[Label]>) -> Vec<Vec<Label>>
     where
-        Query: AsRef<[Label]>, // + 'b
+        Label: Clone,
     {
-        self.inner.common_prefix_search(query)
+        self.inner
+            .common_prefix_search(query)
+            .into_iter()
+            .map(|v| v.into_iter().cloned().collect())
+            .collect()
     }
 
-    /// Return all entries that match `query`.
+    /// Return all entries that match `query`, cloned.
     ///
     /// # Panics
     /// If `query` is empty.
-    pub fn predictive_search(
-        &self,
-        query: impl AsRef<[Label]>,
-    ) -> Defray<SearchIter<'_, Label, ()>> {
-        self.inner.predictive_search(query)
+    pub fn predictive_search(&self, query: impl AsRef<[Label]>) -> Vec<Vec<Label>>
+    where
+        Label: Clone,
+    {
+        let chunk = self.inner.predictive_search(query);
+        chunk.map(|v| v.cloned().collect()).into_iter().collect()
     }
-
-    /// Return the postfixes of all entries that match `query`.
+    /// Return the postfixes of all entries that match `query`, cloned.
     ///
     /// # Panics
     /// If `query` is empty.
-    pub fn postfix_search(&self, query: impl AsRef<[Label]>) -> Defray<PostfixIter<'_, Label, ()>> {
-        self.inner.postfix_search(query)
+    pub fn postfix_search(&self, query: impl AsRef<[Label]>) -> Vec<Vec<Label>>
+    where
+        Label: Clone,
+    {
+        let chunk = self.inner.postfix_search(query);
+        chunk.map(|v| v.cloned().collect()).into_iter().collect()
     }
 
     /// Return true if `query` is a prefix.
@@ -58,13 +66,13 @@ impl<Label: Ord> Trie<Label> {
     }
 }
 
-impl<Label: Ord> Default for TrieBuilder<Label> {
+impl<Label: Ord + Clone> Default for TrieBuilder<Label> {
     fn default() -> Self {
         Self::new()
     }
 }
 
-impl<Label: Ord> TrieBuilder<Label> {
+impl<Label: Ord + Clone> TrieBuilder<Label> {
     pub fn new() -> Self {
         Self {
             inner: map::TrieBuilder::new(),
@@ -94,7 +102,7 @@ impl<Label: Ord> TrieBuilder<Label> {
 
 #[cfg(test)]
 mod search_tests {
-    use crate::trie::{Trie, TrieBuilder};
+    use crate::trie::clone::{Trie, TrieBuilder};
 
     fn build_trie() -> Trie<u8> {
         let mut builder = TrieBuilder::new();
@@ -173,10 +181,7 @@ mod search_tests {
                 fn $name() {
                     let (query, expected_results) = $value;
                     let trie = super::build_trie();
-                    let results: Vec<String> = trie.predictive_search(query)
-                                      .into_iter()
-                                      .map(|g| String::from_utf8(g.into_iter().cloned().collect()).unwrap())
-                                      .collect();
+                    let results = trie.predictive_search(query).into_iter().map(|g| String::from_utf8(g).unwrap()).collect::<Vec<_>>();
                     // results.sort_by(|a, b| a.len().cmp(&b.len()));
                     assert_eq!(results, expected_results);
                 }
@@ -203,11 +208,8 @@ mod search_tests {
                 fn $name() {
                     let (query, expected_results) = $value;
                     let trie = super::build_trie();
-                    let results: Vec<String> = trie.common_prefix_search(query)
-                                      .into_iter()
-                                      .map(|g| String::from_utf8(g.into_iter().cloned().collect()).unwrap())
-                                      .collect();
-                    // let expected_results: Vec<Vec<u8>> = expected_results.iter().map(|s| s.as_bytes().to_vec()).collect();
+                    let results = trie.common_prefix_search(query);
+                    let expected_results: Vec<Vec<u8>> = expected_results.iter().map(|s| s.as_bytes().to_vec()).collect();
                     assert_eq!(results, expected_results);
                 }
             )*
@@ -227,77 +229,6 @@ mod search_tests {
     }
 
     mod posfix_search_tests {
-
-        #[test]
-        fn postfix_unfused() {
-            let trie = super::build_trie();
-            let postfixes = trie.postfix_search("app");
-            let mut iter = postfixes.into_inner().map(|x| *x as char);
-            assert_eq!(iter.next(), Some('p'));
-            assert!(iter.next().is_none());
-            assert_eq!(iter.next(), Some('p'));
-            assert_eq!(iter.next(), Some('l'));
-            assert_eq!(iter.next(), Some('e'));
-        }
-
-        #[test]
-        fn postfix_baseline() {
-            let trie = super::build_trie();
-            let postfixes = trie.postfix_search("app");
-            let mut chunks = postfixes.into_iter();
-            // assert_eq!(chunks.by_ref().count(), 3);
-            let mut iter = chunks.next().unwrap().map(|x| *x as char);
-            assert_eq!(iter.next(), Some('p'));
-            assert_eq!(iter.next(), None);
-            let mut iter = chunks.next().unwrap().map(|x| *x as char);
-            assert_eq!(iter.next(), Some('p'));
-            assert_eq!(iter.next(), Some('l'));
-            assert_eq!(iter.next(), Some('e'));
-            assert_eq!(iter.next(), None);
-            let mut iter = chunks.next().unwrap().map(|x| *x as char);
-            assert_eq!(iter.next(), Some('p'));
-            assert_eq!(iter.next(), Some('l'));
-            assert_eq!(iter.next(), Some('i'));
-            assert_eq!(iter.next(), Some('c'));
-            assert_eq!(iter.next(), Some('a'));
-            assert_eq!(iter.next(), Some('t'));
-            assert_eq!(iter.next(), Some('i'));
-            assert_eq!(iter.next(), Some('o'));
-            assert_eq!(iter.next(), Some('n'));
-            assert_eq!(iter.next(), None);
-
-            assert!(chunks.next().is_none());
-        }
-
-        #[test]
-        fn postfix_2() {
-            let trie = super::build_trie();
-            let postfixes = trie.postfix_search("b");
-            let mut chunks = postfixes.into_iter();
-            let mut iter = chunks.next().unwrap().map(|x| *x as char);
-            assert_eq!(iter.next(), Some('b'));
-            assert_eq!(iter.next(), Some('e'));
-            assert_eq!(iter.next(), Some('t'));
-            assert_eq!(iter.next(), Some('t'));
-            assert_eq!(iter.next(), Some('e'));
-            assert_eq!(iter.next(), Some('r'));
-            assert_eq!(iter.next(), None);
-            assert!(chunks.next().is_none());
-        }
-
-        #[test]
-        fn postfix_3() {
-            let trie = super::build_trie();
-            let postfixes = trie.postfix_search("bet");
-            let mut chunks = postfixes.into_iter();
-            let mut iter = chunks.next().unwrap().map(|x| *x as char);
-            assert_eq!(iter.next(), Some('t'));
-            assert_eq!(iter.next(), Some('t'));
-            assert_eq!(iter.next(), Some('e'));
-            assert_eq!(iter.next(), Some('r'));
-            assert_eq!(iter.next(), None);
-            assert!(chunks.next().is_none());
-        }
 
         #[test]
         fn postfix_no_match() {
